[Refactor] #138 - DiaryBottomView UI 분리
