+++ Smeem-iOS/Smeem-iOS/Presentation/ForeignDiary/DiaryViewController.swift
@@ -10,13 +10,7 @@ import UIKit
 import SnapKit
 
 protocol DiaryStrategy {
-    func configureRandomSubjectButtonImage(_ button: UIButton)
-    func configureToolTipView(_ imageView: UIImageView)
-    func configureRandomSubjectButton(_ button: UIButton)
-    func configureLanguageLabel(_ label: UILabel)
-    func configureLeftNavigationButton(_ button: UIButton)
-    func configureRightNavigationButton(_ button: UIButton)
-    func configureStepLabel(_ label: UILabel)
+    
 }
 
 class DiaryViewController: UIViewController {
@@ -29,8 +23,8 @@ class DiaryViewController: UIViewController {
     
     private var randomTopicEnabled: Bool = false {
         didSet {
-            updateRandomTopicView()
-            updateInputTextViewConstraints()
+            //            updateRandomTopicView()
+            //            updateInputTextViewConstraints()
             view.layoutIfNeeded()
         }
     }
@@ -48,86 +42,15 @@ class DiaryViewController: UIViewController {
     
     // MARK: - UI Property
     
-    let navigationView = UIView()
-    private lazy var randomSubjectView = RandomSubjectView()
-    let loadingView = LoadingView()
-    
-    private let navibarContentStackView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.alignment = .center
-        stackView.distribution = .equalSpacing
-        // FIXME: 기기대응시 문제가 생길수도..?
-        stackView.spacing = 100
-        return stackView
-    }()
-    
-    lazy var leftNavigationButton: UIButton = {
-        let button = UIButton()
-        button.titleLabel?.font = .b4
-        button.setTitleColor(.black, for: .normal)
-        button.contentEdgeInsets = UIEdgeInsets(top: 5, left: 0, bottom: 5, right: 5)
-        button.addTarget(self, action: #selector(leftNavigationButtonDidTap), for: .touchUpInside)
-        return button
-    }()
-    
-    private let languageLabel: UILabel = {
-        let label = UILabel()
-        label.font = .s2
-        label.textColor = .smeemBlack
-        label.text = "Lang"
-        return label
-    }()
-    
-    private let stepLabel: UILabel = {
-        let label = UILabel()
-        label.font = .c4
-        label.textColor = .gray500
-        return label
-    }()
-    
-    lazy var rightNavigationButton: UIButton = {
-        let button = UIButton()
-        button.titleLabel?.font = .b1
-        button.setTitleColor(.gray300, for: .normal)
-        button.setTitle("완료", for: .normal)
-        button.contentEdgeInsets = UIEdgeInsets(top: 5, left: 5, bottom: 5, right: 0)
-        button.addTarget(self, action: #selector(rightNavigationButtonDidTap) , for: .touchUpInside)
-        return button
-    }()
+    let navigationView = SmeemNavigationBar()
     
     let inputTextView = SmeemTextView(type: .editable(SmeemTextViewManager.shared), placeholderText: "일기를 작성해주세요 :)")
     
-    let bottomView: UIView = {
-        let view = UIView()
-        view.backgroundColor = .gray100
-        return view
-    }()
+    let bottomView = DiaryBottomView(viewType: .withHint)
     
-    let thinLine = SeparationLine(height: .thin)
-    
-    lazy var randomSubjectButton: UIButton = {
-        let button = UIButton()
-        button.addTarget(self, action: #selector(randomTopicButtonDidTap), for: .touchUpInside)
-        button.setImage(Constant.Image.btnRandomSubjectInactive, for: .normal)
-        return button
-    }()
-    
-    private lazy var dismissButton: UIButton? = {
-        let button = UIButton()
-        button.addTarget(self, action: #selector(dismissButtonDidTap), for: .touchUpInside)
-        return button
-    }()
-    
-    lazy var randomSubjectToolTip: UIImageView? = {
-        let image = UIImageView()
-        image.image = Constant.Image.icnToolTip
-        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(randomSubjectToolTipDidTap))
-        image.addGestureRecognizer(tapGesture)
-        image.isUserInteractionEnabled = true
-        return image
-    }()
-    
+    private lazy var randomSubjectView = RandomSubjectView()
     var smeemToastView: SmeemToastView?
+    let loadingView = LoadingView()
     
     // MARK: - Life Cycle
     
@@ -141,11 +64,9 @@ class DiaryViewController: UIViewController {
         super.viewDidLoad()
         
         configureDiaryStrategy()
-        configureUI()
         setupUI()
         setDelegate()
-        checkTutorial()
-        checkTooltip()
+//        checkTutorial()
     }
     
     override func viewWillDisappear(_ animated: Bool) {
@@ -156,11 +77,12 @@ class DiaryViewController: UIViewController {
             case is DiaryViewController:
                 print("DiaryViewController")
             case is ForeignDiaryViewController:
-                randomSubjectToolTip?.removeFromSuperview()
+                //                randomSubjectToolTip?.removeFromSuperview()
+                print("ForeignDiaryViewController")
             case is StepOneKoreanDiaryViewController:
                 print("StepOneKoreanDiaryViewController")
             case is StepTwoKoreanDiaryViewController:
-                randomSubjectToolTip?.removeFromSuperview()
+                print("StepTwoKoreanDiaryViewController")
             default:
                 break
             }
@@ -175,85 +97,85 @@ class DiaryViewController: UIViewController {
     
     // MARK: - @objc
     
-    @objc func randomTopicButtonDidTap() {
-        if !UserDefaultsManager.randomSubjectToolTip {
-            UserDefaultsManager.randomSubjectToolTip = true
-            randomSubjectToolTip?.isHidden = true
-        }
-        
-        setRandomTopicButtonToggle()
-        
-        if !isTopicCalled {
-            randomSubjectWithAPI()
-            randomSubjectButton.setImage(Constant.Image.btnRandomSubjectActive, for: .normal)
-            isTopicCalled = true
-        } else {
-            isTopicCalled = false
-            topicID = nil
-        }
-        randomSubjectView.setData(contentText: topicContent)
-    }
+    //    @objc func randomTopicButtonDidTap() {
+    //        if !UserDefaultsManager.randomSubjectToolTip {
+    //            UserDefaultsManager.randomSubjectToolTip = true
+    //            randomSubjectToolTip?.isHidden = true
+    //        }
+    //
+    //        setRandomTopicButtonToggle()
+    //
+    //        if !isTopicCalled {
+    //            randomSubjectWithAPI()
+    //            randomSubjectButton.setImage(Constant.Image.btnRandomSubjectActive, for: .normal)
+    //            isTopicCalled = true
+    //        } else {
+    //            isTopicCalled = false
+    //            topicID = nil
+    //        }
+    //        randomSubjectView.setData(contentText: topicContent)
+    //    }
+    //
+    //    @objc func leftNavigationButtonDidTap() {
+    //        self.navigationController?.popViewController(animated: true)
+    //    }
     
-    @objc func leftNavigationButtonDidTap() {
-        self.navigationController?.popViewController(animated: true)
-    }
+    //    @objc func rightNavigationButtonDidTap() {
+    //        if !rightNavigationButton.isEnabled {
+    //            showToastIfNeeded(toastType: .defaultToast(bodyType: .regEx))
+    //        }
+    //    }
     
-    @objc func rightNavigationButtonDidTap() {
-        if !rightNavigationButton.isEnabled {
-            showToastIfNeeded(toastType: .defaultToast(bodyType: .regEx))
-        }
-    }
-    
-    @objc func keyboardWillShow(notification: NSNotification) {
-        guard let userInfo = notification.userInfo,
-              let keyboardFrame = userInfo[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect
-        else { return }
-        
-        let keyboardHeight = keyboardFrame.height
-        let insets = UIEdgeInsets(top: 0, left: 0, bottom: keyboardHeight, right: 0)
-        inputTextView.contentInset = insets
-        inputTextView.scrollIndicatorInsets = insets
-        self.bottomView.transform = CGAffineTransform(translationX: 0, y: -keyboardHeight)
-        
-        self.keyboardHeight = keyboardFrame.height
-        isKeyboardVisible = true
-        
-        UIView.animate(withDuration: 0.3) {
-            self.bottomView.snp.updateConstraints {
-                $0.height.equalTo(53)
-            }
-            self.view.layoutIfNeeded()
-        }
-        
-        updateAdditionalViewsForKeyboard(notification: notification, keyboardHeight: keyboardHeight)
-        
-    }
-    
-    @objc func keyboardWillHide(notification: NSNotification) {
-        inputTextView.contentInset = .zero
-        inputTextView.scrollIndicatorInsets = .zero
-        self.bottomView.transform = CGAffineTransform.identity
-        
-        isKeyboardVisible = false
-        
-        UIView.animate(withDuration: 0.3) {
-            self.bottomView.snp.updateConstraints {
-                $0.height.equalTo(87)
-            }
-            self.view.layoutIfNeeded()
-        }
-        
-        updateAdditionalViewsForKeyboard(notification: notification, keyboardHeight: 0)
-    }
-    
-    @objc func dismissButtonDidTap() {
-        dismissButton?.removeFromSuperview()
-    }
-    
-    @objc func randomSubjectToolTipDidTap() {
-        self.randomSubjectToolTip?.isHidden = true
-        UserDefaultsManager.randomSubjectToolTip = true
-    }
+    //    @objc func keyboardWillShow(notification: NSNotification) {
+    //        guard let userInfo = notification.userInfo,
+    //              let keyboardFrame = userInfo[UIResponder.keyboardFrameEndUserInfoKey] as? CGRect
+    //        else { return }
+    //
+    //        let keyboardHeight = keyboardFrame.height
+    //        let insets = UIEdgeInsets(top: 0, left: 0, bottom: keyboardHeight, right: 0)
+    //        inputTextView.contentInset = insets
+    //        inputTextView.scrollIndicatorInsets = insets
+    //        self.bottomView.transform = CGAffineTransform(translationX: 0, y: -keyboardHeight)
+    //
+    //        self.keyboardHeight = keyboardFrame.height
+    //        isKeyboardVisible = true
+    //
+    //        UIView.animate(withDuration: 0.3) {
+    //            self.bottomView.snp.updateConstraints {
+    //                $0.height.equalTo(53)
+    //            }
+    //            self.view.layoutIfNeeded()
+    //        }
+    //
+    //        updateAdditionalViewsForKeyboard(notification: notification, keyboardHeight: keyboardHeight)
+    //
+    //    }
+    //
+    //    @objc func keyboardWillHide(notification: NSNotification) {
+    //        inputTextView.contentInset = .zero
+    //        inputTextView.scrollIndicatorInsets = .zero
+    //        self.bottomView.transform = CGAffineTransform.identity
+    //
+    //        isKeyboardVisible = false
+    //
+    //        UIView.animate(withDuration: 0.3) {
+    //            self.bottomView.snp.updateConstraints {
+    //                $0.height.equalTo(87)
+    //            }
+    //            self.view.layoutIfNeeded()
+    //        }
+    //
+    //        updateAdditionalViewsForKeyboard(notification: notification, keyboardHeight: 0)
+    //    }
+    //
+    //    @objc func dismissButtonDidTap() {
+    //        dismissButton?.removeFromSuperview()
+    //    }
+    //
+    //    @objc func randomSubjectToolTipDidTap() {
+    //        self.randomSubjectToolTip?.isHidden = true
+    //        UserDefaultsManager.randomSubjectToolTip = true
+    //    }
     
     // MARK: - Custom Method
     
@@ -281,19 +203,6 @@ class DiaryViewController: UIViewController {
         }
     }
     
-    private func configureUI() {
-        if let randomSubjectToolTip = randomSubjectToolTip {
-            diaryStrategy?.configureToolTipView(randomSubjectToolTip)
-        }
-        
-        diaryStrategy?.configureRandomSubjectButton(randomSubjectButton)
-        diaryStrategy?.configureRandomSubjectButton(randomSubjectButton)
-        diaryStrategy?.configureLanguageLabel(languageLabel)
-        diaryStrategy?.configureLeftNavigationButton(leftNavigationButton)
-        diaryStrategy?.configureRightNavigationButton(rightNavigationButton)
-        diaryStrategy?.configureStepLabel(stepLabel)
-    }
-    
     private func setBackgroundColor() {
         view.backgroundColor = .smeemWhite
     }
@@ -302,33 +211,33 @@ class DiaryViewController: UIViewController {
         randomTopicEnabled.toggle()
     }
     
-    private func updateRandomTopicView() {
-        if randomTopicEnabled {
-            view.addSubview(randomSubjectView)
-            randomSubjectView.snp.makeConstraints {
-                $0.top.equalTo(navigationView.snp.bottom).offset(convertByHeightRatio(16))
-                $0.leading.equalToSuperview()
-            }
-            randomSubjectButton.setImage(Constant.Image.btnRandomSubjectActive, for: .normal)
-        } else {
-            randomSubjectView.removeFromSuperview()
-            randomSubjectButton.setImage(Constant.Image.btnRandomSubjectInactive, for: .normal)
-        }
-    }
-    
-    private func updateInputTextViewConstraints() {
-        inputTextView.snp.remakeConstraints {
-            $0.top.equalTo(randomTopicEnabled ? randomSubjectView.snp.bottom : navigationView.snp.bottom)
-            $0.leading.trailing.equalToSuperview()
-            $0.bottom.equalTo(bottomView.snp.top)
-        }
-    }
+    //    private func updateRandomTopicView() {
+    //        if randomTopicEnabled {
+    //            view.addSubview(randomSubjectView)
+    //            randomSubjectView.snp.makeConstraints {
+    //                $0.top.equalTo(navigationView.snp.bottom).offset(convertByHeightRatio(16))
+    //                $0.leading.equalToSuperview()
+    //            }
+    //            randomSubjectButton.setImage(Constant.Image.btnRandomSubjectActive, for: .normal)
+    //        } else {
+    //            randomSubjectView.removeFromSuperview()
+    //            randomSubjectButton.setImage(Constant.Image.btnRandomSubjectInactive, for: .normal)
+    //        }
+    //    }
+    //
+    //    private func updateInputTextViewConstraints() {
+    //        inputTextView.snp.remakeConstraints {
+    //            $0.top.equalTo(randomTopicEnabled ? randomSubjectView.snp.bottom : navigationView.snp.bottom)
+    //            $0.leading.trailing.equalToSuperview()
+    //            $0.bottom.equalTo(bottomView.snp.top)
+    //        }
+    //    }
     
     func updateAdditionalViewsForKeyboard(notification: NSNotification, keyboardHeight: CGFloat) {}
     
     private func keyboardAddObserver() {
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name: UIResponder.keyboardWillHideNotification, object: nil)
+        //        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
+        //        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name: UIResponder.keyboardWillHideNotification, object: nil)
     }
     
     private func keyboardRemoveObserver() {
@@ -340,28 +249,10 @@ class DiaryViewController: UIViewController {
     
     private func setLayout() {
         view.addSubviews(navigationView, inputTextView, bottomView)
-        navigationView.addSubviews(navibarContentStackView, stepLabel)
-        navibarContentStackView.addArrangedSubviews(leftNavigationButton, languageLabel, rightNavigationButton)
-        bottomView.addSubviews(thinLine, randomSubjectButton)
         
         navigationView.snp.makeConstraints {
             $0.top.leading.trailing.equalTo(view.safeAreaLayoutGuide)
             $0.height.equalTo(convertByHeightRatio(66))
-        }
-        
-        navibarContentStackView.snp.makeConstraints {
-            $0.center.equalToSuperview()
-            if let strategy = diaryStrategy, strategy is ForeignDiaryStrategy || strategy is StepOneKoreanDiaryStrategy {
-                $0.leading.equalToSuperview().offset(convertByWidthRatio(18))
-            } else {
-                $0.leading.equalToSuperview().offset(convertByWidthRatio(12))
-            }
-            $0.height.equalTo(42)
-        }
-        
-        stepLabel.snp.makeConstraints {
-            $0.top.equalTo(languageLabel.snp.bottom).offset(convertByWidthRatio(4))
-            $0.centerX.equalToSuperview()
         }
         
         inputTextView.snp.makeConstraints {
@@ -374,60 +265,30 @@ class DiaryViewController: UIViewController {
             $0.bottom.leading.trailing.equalToSuperview()
             $0.height.equalTo(convertByHeightRatio(87))
         }
-        
-        thinLine.snp.makeConstraints {
-            $0.bottom.equalTo(bottomView.snp.top)
-            $0.centerX.equalToSuperview()
-        }
-        
-        randomSubjectButton.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(convertByHeightRatio(17 - 5))
-            $0.trailing.equalToSuperview().offset(convertByWidthRatio(-18 + 5))
-            $0.width.equalTo(convertByWidthRatio(78 + 10))
-            $0.height.equalTo(convertByHeightRatio(19 + 10))
-        }
-    }
-    
-    private func checkTooltip() {
-        let randomSubjectToolTipe = UserDefaultsManager.randomSubjectToolTip
-        
-        if !randomSubjectToolTipe {
-            
-            view.addSubview(randomSubjectToolTip ?? UIImageView())
-            
-            randomSubjectToolTip?.snp.makeConstraints {
-                $0.width.equalTo(convertByWidthRatio(180))
-                $0.height.equalTo(convertByHeightRatio(48))
-                $0.bottom.equalTo(view.keyboardLayoutGuide.snp.top).offset(constraintByNotch(-37, -42))
-                $0.trailing.equalToSuperview().inset(convertByHeightRatio(18))
-            }
-        } else {
-            randomSubjectToolTip = nil
-        }
     }
     
     private func checkTutorial() {
-        //        if self is StepOneKoreanDiaryViewController {
-        //            let tutorialDiaryStepOne = UserDefaultsManager.tutorialDiaryStepOne
-        //
-        //            if !tutorialDiaryStepOne {
-        //                UserDefaultsManager.tutorialDiaryStepOne = true
-        //
-        //                view.addSubviews(tutorialImageView ?? UIImageView(), dismissButton ?? UIButton())
-        //
-        //                tutorialImageView?.snp.makeConstraints {
-        //                    $0.top.leading.trailing.bottom.equalToSuperview()
-        //                }
-        //                dismissButton?.snp.makeConstraints {
-        //                    $0.top.equalToSuperview().inset(convertByHeightRatio(204))
-        //                    $0.trailing.equalToSuperview().inset(convertByHeightRatio(10))
-        //                    $0.width.height.equalTo(convertByHeightRatio(45))
-        //                }
-        //            } else {
-        //                tutorialImageView = nil
-        //                dismissButton = nil
-        //            }
-        //        }
+//        if self is StepOneKoreanDiaryViewController {
+//            let tutorialDiaryStepOne = UserDefaultsManager.tutorialDiaryStepOne
+//
+//            if !tutorialDiaryStepOne {
+//                UserDefaultsManager.tutorialDiaryStepOne = true
+//
+//                view.addSubviews(tutorialImageView ?? UIImageView(), dismissButton ?? UIButton())
+//
+//                tutorialImageView?.snp.makeConstraints {
+//                    $0.top.leading.trailing.bottom.equalToSuperview()
+//                }
+//                dismissButton?.snp.makeConstraints {
+//                    $0.top.equalToSuperview().inset(convertByHeightRatio(204))
+//                    $0.trailing.equalToSuperview().inset(convertByHeightRatio(10))
+//                    $0.width.height.equalTo(convertByHeightRatio(45))
+//                }
+//            } else {
+//                tutorialImageView = nil
+//                dismissButton = nil
+//            }
+//        }
     }
 }
 
@@ -485,7 +346,7 @@ extension DiaryViewController {
                 let homeVC = HomeViewController()
                 homeVC.toastMessageFlag = true
                 homeVC.badgePopupData = self.badgePopupContent
-                self.randomSubjectToolTip = nil
+                //                self.randomSubjectToolTip = nil
                 let rootVC = UINavigationController(rootViewController: homeVC)
                 self.changeRootViewControllerAndPresent(rootVC)
             }
